Cut auto folding of python files
Folding would work except for instances where parameters, or really
anything that is indented more that four spaces counts as multiple
indents.  This is not really what I was expecting. So a line like

  if does_really.long_function_name(parameter_one(argument1, argument2),
                                    parameter_two,
                                    parameter_three):

would wrap the seconds two parameter lines.  However, this is not what I
would think of as a send level of indent.  That really should only be
used for complete statements.